refactor(gradle)!: Split out authentication utilities to a new module
This allows for more fine-granular dependency management.

BREAKING CHANGE: Programmatic users might need to depend on
`authorization-utils` instead of / in addition to `ort-utils` now.
