+++ plugins/version-control-systems/git/build.gradle.kts
@@ -39,6 +39,7 @@ dependencies {
     implementation(libs.kotlinx.serialization.core)
     implementation(libs.kotlinx.serialization.xml)
     implementation(libs.semver4j)
+    implementation(projects.utils.authenticationUtils)
     implementation(projects.utils.commonUtils)
     implementation(projects.utils.ortUtils)
