+++ utils/authentication/src/main/kotlin/OrtProxySelector.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.utils.ort
+package org.ossreviewtoolkit.utils.authentication
 
 import java.io.IOException
 import java.net.InetSocketAddress
